Reorganize sources into sensible hierarchy
