+++ blobxfer/operations/azure/blob/block.py
@@ -23,28 +23,45 @@
 # DEALINGS IN THE SOFTWARE.
 
 # compat imports
-from __future__ import absolute_import, division, print_function  # noqa
+from __future__ import absolute_import, division, print_function
 from builtins import (  # noqa
     bytes, dict, int, list, object, range, ascii, chr, hex, input,
     next, oct, open, pow, round, super, filter, map, zip
 )
 # stdlib imports
+import logging
 # non-stdlib imports
+import azure.storage.blob
 # local imports
+import blobxfer.retry
 
-from .operations.azure.blob.append import (  # noqa
-    create_client as create_append_blob_client
-)
-from .operations.azure.blob.block import (  # noqa
-    create_client as create_block_blob_client
-)
-from .operations.azure.blob.page import (  # noqa
-    create_client as create_page_blob_client
-)
-from .operations.azure.file import (  # noqa
-    create_client as create_file_client
-)
+# create logger
+logger = logging.getLogger(__name__)
 
-from .operations.download import (  # noqa
-    Downloader
-)
+
+def create_client(storage_account):
+    # type: (blobxfer.models.azure.StorageAccount) -> BlockBlobService
+    """Create block blob client
+    :param blobxfer.models.azure.StorageAccount storage_account:
+        storage account
+    :rtype: azure.storage.blob.BlockBlobService
+    :return: block blob service client
+    """
+    if storage_account.is_sas:
+        client = azure.storage.blob.BlockBlobService(
+            account_name=storage_account.name,
+            sas_token=storage_account.key,
+            endpoint_suffix=storage_account.endpoint)
+    else:
+        client = azure.storage.blob.BlockBlobService(
+            account_name=storage_account.name,
+            account_key=storage_account.key,
+            endpoint_suffix=storage_account.endpoint)
+    # set retry policy
+    client.retry = blobxfer.retry.ExponentialRetryWithMaxWait().retry
+    return client
+
+
+def upload_block():
+    logger.info('upload block')
+    print('upload')
